DI is sixteen bits, and the row steppers are DI
cga_next_row, cga_prev_row and their two `ja` variants took and returned
uint32_t while modelling one 16-bit register. They are uint16_t now, so
the wrap at the end of the odd half is the original's rather than
something C is trusted to reach by a different route.

This one is not free: .text grows 232 bytes, because the truncation that
was implicit in the type is now emitted. That is the first change in this
sweep that the compiler notices, so it is checked the way a change to
behaviour has to be rather than by comparing bytes - the eight routines
that step rows for a living all agree with the original 10 calls each,
and 2,000 frames on level 11, 1,500 on level 50 and 1,500 through the
end-of-level bonus are identical throughout.

+++ src/game.h
@@ -1124,8 +1124,11 @@ uint8_t *exepack_load(const char *path, size_t *out_len);
 
 extern uint8_t g_vram[CGA_SIZE];
 
-/* Step a CGA offset on by one scan line, the way the game does. */
-static inline uint32_t cga_next_row(uint32_t di)
+/* Step a CGA offset on by one scan line, the way the game does. DI is a
+ * 16-bit register and these are that register, so they are that wide -
+ * which is also what makes the wrap at the end of the odd half the
+ * original's rather than something C would do differently. */
+static inline uint16_t cga_next_row(uint16_t di)
 {
     return di < CGA_PLANE ? di + CGA_PLANE : di - (CGA_PLANE - CGA_STRIDE);
 }
@@ -1140,19 +1143,19 @@ static inline uint32_t cga_next_row(uint32_t di)
  * Getting this wrong sends an offset of 0x2000 to 0x3fb0, past the bottom of
  * the visible screen into the padding at the end of the plane. That is what
  * the frame's scroll was doing. */
-static inline uint32_t cga_prev_row(uint32_t di)
+static inline uint16_t cga_prev_row(uint16_t di)
 {
     return di >= CGA_PLANE ? di - CGA_PLANE : di + (CGA_PLANE - CGA_STRIDE);
 }
 
 /* intro_logo's pair: `ja`, so 0x2000 goes the other way. Only 1ac2:54d6 uses
  * these, at 54f6, 5535, 557a and 55b6. */
-static inline uint32_t cga_prev_row_ja(uint32_t di)
+static inline uint16_t cga_prev_row_ja(uint16_t di)
 {
     return di > CGA_PLANE ? di - CGA_PLANE : di + (CGA_PLANE - CGA_STRIDE);
 }
 
-static inline uint32_t cga_next_row_ja(uint32_t di)
+static inline uint16_t cga_next_row_ja(uint16_t di)
 {
     return di > CGA_PLANE ? di - (CGA_PLANE - CGA_STRIDE) : di + CGA_PLANE;
 }
